Put grammar in a better place. Separate the expression code into a namespace.
