+++ distribution/lib/Standard/Table/0.0.0-dev/src/Data/Expression.enso
@@ -2,7 +2,7 @@ from Standard.Base import all
 
 import project.Data.Column.Column
 
-polyglot java import org.enso.table.parsing.ExpressionEvaluator
+polyglot java import org.enso.table.expressions.ExpressionEvaluator
 
 type Expression
     For_Column expression:Text
